Move GSICAPBENCH tests into separate file; add __init__.py to test directory to allow imports
