+++ src/test/test.py
@@ -11,12 +11,11 @@ import subprocess
 import tempfile
 import time
 import unittest
-from shutil import copy, copytree
+from shutil import copy
 from string import Template
 
 import djerba.util.constants as constants
 import djerba.util.ini_fields as ini
-from djerba.benchmark import benchmarker
 from djerba.configure import configurer, log_r_cutoff_finder
 from djerba.extract.extractor import extractor
 from djerba.extract.r_script_wrapper import r_script_wrapper
@@ -126,55 +125,6 @@ class TestArchive(TestBase):
             data = json.loads(archive_file.read())
         self.assertEqual(len(data['report']), 19)
         self.assertEqual(len(data['supplementary']['config']), 3)
-
-class TestBenchmark(TestBase):
-
-    class mock_args_compare:
-        """Use instead of argparse to store params for testing"""
-
-        def __init__(self, report_dirs, compare_all=False):
-            self.subparser_name = constants.COMPARE
-            self.report_dir = report_dirs
-            self.compare_all = compare_all
-            # logging
-            self.log_path = None
-            self.debug = False
-            self.verbose = False
-            self.quiet = True
-
-    class mock_args_report:
-        """Use instead of argparse to store params for testing"""
-
-        def __init__(self, input_dir, output_dir):
-            self.subparser_name = constants.REPORT
-            self.input_dir = input_dir
-            self.output_dir = output_dir
-            self.dry_run = False
-            # logging
-            self.log_path = None
-            self.debug = False
-            self.verbose = False
-            self.quiet = True
-
-    def test_benchmark(self):
-        out_dir = self.tmp_dir
-        input_dir = os.path.join(self.bench_dir, 'benchmark')
-        report_dir = os.path.join(out_dir, 'report')
-        os.mkdir(report_dir)
-        report_args = self.mock_args_report(input_dir, report_dir)
-        self.assertTrue(benchmarker(report_args).run())
-        report_1a = os.path.join(report_dir, 'GSICAPBENCH_1219')
-        report_1b = os.path.join(report_dir, 'GSICAPBENCH_1219.copy')
-        copytree(report_1a, report_1b) # make a copy to test identical inputs
-        report_2 = os.path.join(report_dir, 'GSICAPBENCH_1232')
-        compare_args_1 = self.mock_args_compare([report_1a, report_1b])
-        self.assertTrue(benchmarker(compare_args_1).run())
-        compare_args_2 = self.mock_args_compare([report_1a, report_2])
-        self.assertFalse(benchmarker(compare_args_2).run())
-        compare_args_3 = self.mock_args_compare([report_1a, report_1b], compare_all=True)
-        self.assertTrue(benchmarker(compare_args_3).run())
-        compare_args_4 = self.mock_args_compare([report_1a, report_2], compare_all=True)
-        self.assertFalse(benchmarker(compare_args_4).run())
 
 class TestConfigure(TestBase):
